final touches really now

+++ WPF/ViewModels/GuideViewModels/InfoWindowViewModel.cs
@@ -66,34 +66,39 @@ namespace BookingApp.WPF.ViewModels.GuideViewModels
 
         internal void quit_Click()
         {
-            List<TourStartTime> dates = TourStartTimeService.GetInstance().GetAll();
-            List<TouristReservation> reservations = TourReservationService.GetInstance().GetAll();
-            foreach (TourStartTime date in dates)
+            MessageBoxResult result = MessageBox.Show("Are you sure you want to delete your account?", "Confirmation", MessageBoxButton.YesNo, MessageBoxImage.Question);
+
+            if (result == MessageBoxResult.Yes)
             {
-                Tour tour = TourService.GetInstance().GetById(date.TourId);
-                if (tour.OwnerId == user.Id && date.Status == "scheduled")
+                List<TourStartTime> dates = TourStartTimeService.GetInstance().GetAll();
+                List<TouristReservation> reservations = TourReservationService.GetInstance().GetAll();
+                foreach (TourStartTime date in dates)
                 {
-                    date.Status = "canceled";
-                    TourStartTimeService.GetInstance().Update(date);
-
-                    foreach(TouristReservation reservation in reservations)
+                    Tour tour = TourService.GetInstance().GetById(date.TourId);
+                    if (tour.OwnerId == user.Id && date.Status == "scheduled")
                     {
-                        if(reservation.Id_TourTime == date.Id)
-                        {
-                           TourVoucher voucher = new TourVoucher();
-                            voucher.TouristId = reservation.Id_Tourist;
-                            voucher.ExpirationDate = DateTime.Now.AddDays(365);
-                            TourVoucherService.GetInstance().Add(voucher);
+                        date.Status = "canceled";
+                        TourStartTimeService.GetInstance().Update(date);
 
+                        foreach (TouristReservation reservation in reservations)
+                        {
+                            if (reservation.Id_TourTime == date.Id)
+                            {
+                                TourVoucher voucher = new TourVoucher();
+                                voucher.TouristId = reservation.Id_Tourist;
+                                voucher.ExpirationDate = DateTime.Now.AddDays(365);
+                                TourVoucherService.GetInstance().Add(voucher);
+
+                            }
                         }
                     }
-                }
-                
-            }
-            GuideService.GetInstance().DeleteGuide(user.Id);
-            quitEvent?.Invoke(this, EventArgs.Empty);
-            infoWindow.Close();
 
+                }
+                GuideService.GetInstance().DeleteGuide(user.Id);
+                quitEvent?.Invoke(this, EventArgs.Empty);
+                infoWindow.Close();
+
+            }
         }
 
         internal void GeneratePdfReport_Click(DateTime? startDate, DateTime? endDate)
